validate event_name is not blank only create ticket

+++ app/controllers/tickets_controller.rb
@@ -34,13 +34,17 @@ class TicketsController < ApplicationController
     param = params[:params]
     @ticket = Ticket.create(ticket_params)
 
-    @ticket.categories << Category.find_by(id: param[:category_id])
-
-    param[:tags].to_s.split(",").each do |name|
-      @ticket.tags <<  Tag.find_or_create_by(name: name.strip)
-    end
+#    @ticket.categories << Category.find_by(id: param[:category_id])
     
-    redirect_to(@ticket, notice: t('success_message'))
+    if @ticket.valid?
+      param[:tags].to_s.split(",").each do |name|
+        @ticket.tags <<  Tag.find_or_create_by(name: name.strip)
+      end
+      
+      redirect_to(@ticket, notice: t('success_message'))
+    else
+      redirect_to :back
+    end
     
   end
 
@@ -52,19 +56,24 @@ class TicketsController < ApplicationController
   # PATCH/PUT /tickets/1
   # PATCH/PUT /tickets/1.json
   def update
-    respond_to do |format|
-      TicketTag.delete_all("ticket_id = '#{@ticket.id}'")
-      params[:params][:tags].to_s.split(",").each do |name|
-        @ticket.tags <<  Tag.find_or_create_by(name: name.strip)
+    if @ticket.valid?
+      respond_to do |format|
+        TicketTag.delete_all("ticket_id = '#{@ticket.id}'")
+        params[:params][:tags].to_s.split(",").each do |name|
+          @ticket.tags <<  Tag.find_or_create_by(name: name.strip)
+        end
+        if @ticket.update(ticket_params)
+          format.html { redirect_to @ticket, notice: 'Ticket was successfully updated.' }
+          format.json { render :show, status: :ok, location: @ticket }
+        else
+          format.html { render :edit }
+          format.json { render json: @ticket.errors, status: :unprocessable_entity }
+        end
       end
-      if @ticket.update(ticket_params)
-        format.html { redirect_to @ticket, notice: 'Ticket was successfully updated.' }
-        format.json { render :show, status: :ok, location: @ticket }
-      else
-        format.html { render :edit }
-        format.json { render json: @ticket.errors, status: :unprocessable_entity }
-      end
+    else
+      redirect_to :back
     end
+    
   end
 
   # DELETE /tickets/1
